Add function "natural_keys" in  module "sorting_tools" that using as sort key for columns header lists.
