+++ parsers/json_files_parser.py
@@ -1,18 +1,5 @@
 import json
-import re
-
-
-def atoi(text):
-    return int(text) if text.isdigit() else text
-
-
-def natural_keys(text):
-    '''
-    alist.sort(key=natural_keys) sorts in human order
-    http://nedbatchelder.com/blog/200712/human_sorting.html
-    (See Toothy's implementation in the comments)
-    '''
-    return [atoi(c) for c in re.split(r'(\d+)', text)]
+from parsers.sorting_tools import natural_keys
 
 
 class JsonFilesParser:
